Caculo/smart-farming-platform/requirements.txt
streamlit>=1.28.0
tensorflow-cpu>=2.15.0
Pillow>=9.0.0
requests>=2.25.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.5.0